# Speedup internal tests