--- a/.gitignore
+++ b/.gitignore
@@ -39,6 +39,7 @@ jspm_packages/
 # Typescript v1 declaration files
 typings/
 !tests/typings
+tests/typings/**/dist
 
 # Optional npm cache directory
 .npm