--- a/package.json
+++ b/package.json
@@ -29,7 +29,7 @@
     "test": "run-s test:*",
     "test:eslint": "eslint --cache packages examples scripts tests",
     "test:typings": "npm run generate:typings && node tests/typings/copy && npm run ts && npm run clean:tests",
-    "test:coverage": "jest --coverage --maxWorkers=8",
+    "test:coverage": "jest --coverage",
     "test:smoke": "babel-node ./tests/smoke.runner.js",
     "ts": "run-p ts:*",
     "ts:sync": "cd tests/typings/sync && tsc --incremental",