--- a/tests/smoke.runner.js
+++ b/tests/smoke.runner.js
@@ -5,29 +5,24 @@ import assert from 'assert'
 import launch from './helpers/launch'
 import { SERVICE_LOGS, LAUNCHER_LOGS, REPORTER_LOGS } from './helpers/fixtures'
 
-const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
+const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms))
 
-(async () => {
-    /**
-     * Mocha wdio testrunner tests
-     */
+const mochaTestrunner = async () => {
     await launch(
         path.resolve(__dirname, 'helpers', 'config.js'),
         { specs: [path.resolve(__dirname, 'mocha', 'test.js')] })
+}
 
-    /**
-     * Jasmine wdio testrunner tests
-     */
+const jasmineTestrunner = async () => {
     await launch(
         path.resolve(__dirname, 'helpers', 'config.js'),
         {
             specs: [path.resolve(__dirname, 'jasmine', 'test.js')],
             framework: 'jasmine'
         })
+}
 
-    /**
-     * Cucumber wdio testrunner tests
-     */
+const cucumberTestrunner = async () => {
     await launch(
         path.resolve(__dirname, 'helpers', 'config.js'),
         {
@@ -38,10 +33,9 @@ const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
             }
         }
     )
+}
 
-    /**
-     * Cucumber fail due to failAmbiguousDefinitions
-     */
+const cucumberFailAmbiguousDefinitions = async () => {
     const hasFailed = await launch(
         path.resolve(__dirname, 'helpers', 'config.js'),
         {
@@ -57,10 +51,9 @@ const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
         () => true
     )
     assert.equal(hasFailed, true)
+}
 
-    /**
-     * wdio test run with custom service
-     */
+const customService = async () => {
     await launch(
         path.resolve(__dirname, 'helpers', 'config.js'),
         {
@@ -72,10 +65,9 @@ const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
     assert.equal(serviceLogs.toString(), SERVICE_LOGS)
     const launcherLogs = fs.readFileSync(path.join(__dirname, 'helpers', 'launcher.log'))
     assert.equal(launcherLogs.toString(), LAUNCHER_LOGS)
+}
 
-    /**
-     * wdio test run with custom reporter as string
-     */
+const customReporterString = async () => {
     await launch(
         path.resolve(__dirname, 'helpers', 'config.js'),
         {
@@ -87,19 +79,17 @@ const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
     const reporterLogs = fs.readFileSync(reporterLogsPath)
     assert.equal(reporterLogs.toString(), REPORTER_LOGS)
     fs.unlinkSync(reporterLogsPath)
+}
 
-    /**
-     * wdio test run with custom reporter as object
-     */
+const customReporterObject = async () => {
     await launch(path.resolve(__dirname, 'helpers', 'reporter.conf.js'), {})
     const reporterLogsWithReporterAsObjectPath = path.join(__dirname, 'helpers', 'wdio-0-0-CustomSmokeTestReporter-reporter.log')
     const reporterLogsWithReporterAsObject = fs.readFileSync(reporterLogsWithReporterAsObjectPath)
     assert.equal(reporterLogsWithReporterAsObject, REPORTER_LOGS)
     fs.unlinkSync(reporterLogsWithReporterAsObjectPath)
+}
 
-    /**
-     * multiremote wdio testrunner tests
-     */
+const multiremote = async () => {
     await launch(
         path.resolve(__dirname, 'helpers', 'config.js'),
         {
@@ -114,10 +104,9 @@ const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
             }
         }
     )
+}
 
-    /**
-     * specfile-level retries
-     */
+const retryFail = async () => {
     const retryFailed = await launch(
         path.resolve(__dirname, 'helpers', 'config.js'),
         {
@@ -129,7 +118,9 @@ const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
         () => true
     )
     assert.equal(retryFailed, true, 'Expected retries to fail but they passed')
+}
 
+const retryPass = async () => {
     let retryFilename = path.join(__dirname, '.retry_succeeded')
     let logfiles = ['wdio-0-0.log', 'wdio-0-1.log'].map(f => path.join(__dirname, f))
     let rmfiles = [retryFilename, ...logfiles]
@@ -156,6 +147,22 @@ const sleep = (ms) => new Promise(resolve => setTimeout(resolve, ms));
     if (fs.existsSync(logfiles[1])) {
         throw Error(`Expected ${logfiles[1]} to not exist but it does`)
     }
+}
+
+(async () => {
+    const tests = [
+        mochaTestrunner(),
+        jasmineTestrunner(),
+        cucumberTestrunner(),
+        cucumberFailAmbiguousDefinitions(),
+        customService(),
+        customReporterString(),
+        customReporterObject(),
+        multiremote(),
+        retryFail(),
+        retryPass(),
+    ]
+    await Promise.all(tests)
 
     /**
      * for some reason the process get stuck therefor exit it