--- a/tests/typings/sync/tsconfig.json
+++ b/tests/typings/sync/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "outDir": "dist",
     "noImplicitAny": true,
     "types": [
       "@wdio/sync",