--- a/tests/typings/webdriverio-jasmine/tsconfig.json
+++ b/tests/typings/webdriverio-jasmine/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "outDir": "dist",
     "noImplicitAny": true,
     "types": [
       "webdriverio",