--- a/tests/typings/webdriverio-mocha/tsconfig.json
+++ b/tests/typings/webdriverio-mocha/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "outDir": "dist",
     "noImplicitAny": true,
     "types": [
       "webdriverio",