## Proposed changes

Reduce time taken by tests execution. Results first, ofcourse!
```
before: 1m 20 sec
after:  0m 33 sec
```

- run jest tests in parallel by removing [detectopenhandles](https://jestjs.io/docs/en/cli#detectopenhandles), 40s -> 14s
- speedup eslint _locally_ by using cache, 9s -> 1s
- speedup types test _locally_ by using --incremental, 7s -> 4s
- run smoke tests in parallel, 22s -> 10s

Also removed comments from generated files

## Types of changes

- [x] internal

## Checklist

- [x] I have read the [CONTRIBUTING](https://github.com/webdriverio/webdriverio/blob/master/CONTRIBUTING.md) doc
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added necessary documentation (if appropriate)

## Further comments


### Reviewers: @webdriverio/technical-committee