@mgrybyk 2019-07-31
@christian-bromann one second, I wanted to run jest in parallel on travis CI as well